requirements.txt
flask
miv-os
requests
numpy
matplotlib
